separate menu example code into new sample mod
