+++ OWML.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 18
-VisualStudioVersion = 18.3.11527.330 d18.3
+VisualStudioVersion = 18.3.11527.330
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{CEAC0A30-FF6F-4618-9C55-E2A0D81B863A}"
 EndProject
@@ -56,6 +56,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OWML.EnumExample", "src\SampleMods\OWML.EnumExample\OWML.EnumExample.csproj", "{739D16FB-7848-4047-A173-500CE7C40399}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OWML.MenuExample", "src\SampleMods\OWML.MenuExample\OWML.MenuExample.csproj", "{16B9B21F-1866-4487-8D03-94CFE7E9DCE5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -139,6 +141,10 @@ Global
 		{739D16FB-7848-4047-A173-500CE7C40399}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{739D16FB-7848-4047-A173-500CE7C40399}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{739D16FB-7848-4047-A173-500CE7C40399}.Release|Any CPU.Build.0 = Release|Any CPU
+		{16B9B21F-1866-4487-8D03-94CFE7E9DCE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{16B9B21F-1866-4487-8D03-94CFE7E9DCE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{16B9B21F-1866-4487-8D03-94CFE7E9DCE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{16B9B21F-1866-4487-8D03-94CFE7E9DCE5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -164,9 +170,10 @@ Global
 		{2E866A8D-1C5B-43C2-80D1-632E785A6F3B} = {CEAC0A30-FF6F-4618-9C55-E2A0D81B863A}
 		{F56A9B0D-420F-493F-9B53-0CCD218A895E} = {C447A599-2700-44E1-BBFA-52880B7BFFBA}
 		{739D16FB-7848-4047-A173-500CE7C40399} = {C447A599-2700-44E1-BBFA-52880B7BFFBA}
+		{16B9B21F-1866-4487-8D03-94CFE7E9DCE5} = {C447A599-2700-44E1-BBFA-52880B7BFFBA}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		EnterpriseLibraryConfigurationToolBinariesPath = packages\Unity.2.1.505.2\lib\NET35;packages\Unity.2.1.505.0\lib\NET35
 		SolutionGuid = {0E767163-75F9-420A-80EB-320429543CAD}
+		EnterpriseLibraryConfigurationToolBinariesPath = packages\Unity.2.1.505.2\lib\NET35;packages\Unity.2.1.505.0\lib\NET35
 	EndGlobalSection
 EndGlobal
